+++ internal/cli/root.go
@@ -95,7 +95,7 @@ func init() {
 		}
 	})
 
-	rootCmd.AddCommand(serveCmd, daemonCmd, stopCmd, restartCmd, initCmd, setupCmd, versionCmd, configCmd, sessionsCmd, statsCmd, diagnosticsCmd, doctorCmd, logLevelCmd)
+	rootCmd.AddCommand(serveCmd, daemonCmd, stopCmd, restartCmd, initCmd, setupCmd, versionCmd, configCmd, sessionsCmd, statsCmd, diagnosticsCmd, doctorCmd, logLevelCmd, debugCmd)
 }
 
 func availableCommandNameWidth(cmd *cobra.Command) int {